feat: change all distros to share the same playbook
This commit replaces all current distro specific playbooks
with symbolic links to a new common playbook.  Further
refactoring is expected as we pivot to a common flow that
they all would share, with minor changes as needed per
distro, rather than the current case where each distro
has it's only copy of the entire deployment which is for
the most part the same for all distros.  Also fixed a
broken symlink for the eks host group vars.