feat: normalize inventory groups and cleanup
This commit remove the distro prefix from the group names in
the inventory files so that the names are consistent and are
common across distros.  Additional cleanup to refactor tasks
like systemd user services to be shared across service roles,
moving certificate info to common config and moving tasks to
check that nodes are up into the tasks which create them.