feat: add support for service node in container to okd
This commit allows for the service node in the inventory to be
assigned type=container and a podman pod and containers will
replace the guest vm for the service node.  Numerous other
fixes and reorgs as well, including the creation of new tasks
to consolidate the creation of credentials used to secure
services.