feat: reorganize the prepare_services tasks of each distro into kubeinit_services role
This commit (and the previous two containing simple deletes and renames) combines all of
the prepare_services tasks across all the distros into a common path located in the
kubeinit_services role.  Other changes include moving the OCP release level from 4.7.1
to 4.7.18 and updating the RHCOS release from 4.6.1 to 4.7.13 to match.