feat: cleanup use of convenience facts
This commit defines some new convenience facts for performing
some of the more common tests in when expressions.  Other changes
include simplifying the collection of hostvars and facts to be
within the same (main) play in the playbook to eventually enable
the ability to run --start-from-task to skip work in the main
play that does not need to be repeated.