fix: add disk full path deploying coreos as guest
Currently defining the VM on the deploy_coreos_guest task does
not add the path of the qcow2 disk. This makes the variable
kubeinit_libvirt_target_image_dir useless.

This patch add the path to the disk in the same way as is defined
on the deploy_centos_guest task.